Moved IconButton into it's own folder
