+++ lib/Button/index.js
@@ -1,2 +1,1 @@
 export { default } from './Button';
-export { default as IconButton } from './IconButton';
