+++ lib/Pane/stories/PaneHeader.js
@@ -6,7 +6,7 @@ import React from 'react';
 import Paneset from '../../Paneset';
 import PaneMenu from '../../PaneMenu';
 import Pane from '../Pane';
-import { Button, IconButton } from '../../Button';
+import IconButton from '../../IconButton';
 
 const firstMenu = (
   <PaneMenu>
