+++ lib/PaneHeader/PaneHeader.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import css from './PaneHeader.css';
-import { IconButton } from '../Button';
+import IconButton from '../IconButton';
 
 export default class PaneHeader extends Component {
   static propTypes = {
